Feat: 게시글 및 댓글 신고 기능 구현

+++ backend/src/main/java/com/gomin_jungdok/gdgoc/comment/dto/CommentCreateDTO.java
@@ -6,6 +6,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public class CommentCreateDTO {
-    private Integer parentCommentId;
+    private Long parentCommentId;
     private String description;
 }
